Rename WSID to WebSocketID
