+++ Sources/WS/Extensions/Request+Observer.swift
@@ -5,5 +5,5 @@ extension Request {
     public func ws() -> AnyObserver { application.ws.observer() }
     
     /// Selected websocket observer
-    public func ws<Observer>(_ wsid: WSID<Observer>) -> Observer { application.ws.observer(wsid) }
+    public func ws<Observer>(_ wsid: WebSocketID<Observer>) -> Observer { application.ws.observer(wsid) }
 }
